Comply with the final exercise's requirements
There's might be some mistakes in the requirements and the solution of this exercise : there is no class
".body", it should be "body" to target the body in the solution.css file. In the requirements, it is said
that the sidebar can't be smaller than "300px" whereas I can't find it specified in the solution.css.
Seems like its the cards themselves that have to "be" 300px wide.